switcher-playground-python/requirements.txt
switcher-client==0.1.1
python-dotenv==1.2.2